Update prompt_manager gem to version 0.5.0 and enhance README
This commit updates the `prompt_manager` gem from version 0.4.2 to 0.5.0,
ensuring we are using the latest features and fixes. In addition, the
README file has been enhanced to improve navigation with new links in
the Table of Contents. The descriptions for certain sections have also
been clarified to provide better context and usability for users.